requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
httpx
razorpay
pymongo[srv]
motor
redis[hiredis]
slowapi
